% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_trendfilter.R
\name{cv_trendfilter}
\alias{cv_trendfilter}
\title{Leave-vth-out cross validation for trendfilter}
\usage{
cv_trendfilter(
  y,
  x = seq_along(y),
  weights = rep(1, n),
  k = 2L,
  error_measure = c("deviance", "mse", "mae"),
  nfolds = 5L,
  ...
)
}
\arguments{
\item{y}{vector of observations of length `n`}

\item{x}{vector of positions at which the `y` have been observed, defaults
to `1:n`}

\item{weights}{vector of weights for the observations, defaults to `rep(1, n)`}

\item{k}{Integer. Degree of the piecewise polynomial curve to be
estimated. For example, `korder = 0` corresponds to a piecewise constant
curve.}

\item{error_measure}{Metric used to calculate cross validation scores. May
be `mse`, `mae`, or `deviance`.}

\item{nfolds}{Integer. The number of folds to use. For leave-vth-out cross
validation, every vth `y` value and its corresponding position (and weight)
are placed into the same fold. The first and last observations are not
assigned to any folds. This value must be at least 2. As an example, with
15 data points and `nfolds = 4`, the points are assigned to folds in the
following way:
\deqn{
0 \; 1 \; 2 \; 3 \; 4 \; 1 \; 2 \; 3 \;  4 \; 1 \; 2 \; 3 \; 4 \; 1 \; 0
}{0 1 2 3 4 1 2 3 4 1 2 3 4 1 0} where 0 indicates no assignment.
Therefore, the folds are not random and running `cv_trendfilter()` twice
will give the same result.}

\item{...}{additional parameters passet along to `trendfilter()`.}
}
\value{
An object with S3 class `"cv_trendfilter"`. Among the list components:
* `full_fit` An object with S3 class `"trendfilter"`, estimated with all
  `y` and `lambda`
* `cv_scores` leave-vth-out cross validation scores
* `cv_se` leave-vth-out cross validation standard error
* `lambda_min` lambda which achieved the optimal cross validation score
* `lambda_1se` lambda that gives the optimal cross validation score
within one standard error.
* `lambda` the value of `lambda` used in the algorithm.
}
\description{
Leave-vth-out cross validation for trendfilter
}
\examples{
x <- 1:100 / 101 * 2 * pi
y <- sin(x) + .2 * rnorm(100)
cv <- cv_trendfilter(y, x, nlambda = 20L)
}
