% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-lists.R
\name{trendfilter_control_list}
\alias{trendfilter_control_list}
\title{Trendfilter control list}
\usage{
trendfilter_control_list(
  obj_tol = 1e-06,
  x_cond = 1e+11,
  admm_control = admm_control_list(),
  ...
)
}
\arguments{
\item{obj_tol}{Double. The convergence tolerance for the objective function.}

\item{x_cond}{Double. The conditioning of the \code{x} values}

\item{admm_control}{An object of class \code{admm_control} used to control
as created by \code{\link[=admm_control_list]{admm_control_list()}}}

\item{...}{not used}
}
\value{
An object of class \code{trendfilter_control}
}
\description{
These arguments are used to handle the internals of \code{\link[=trendfilter]{trendfilter()}}
}
\examples{
trendfilter_control_list()
trendfilter_control_list(obj_tol = 1e-12)
trendfilter_control_list(admm_control = admm_control_list(tolerance = 1e-5))
}
