% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control-lists.R
\name{admm_control_list}
\alias{admm_control_list}
\title{ADMM control list}
\usage{
admm_control_list(
  max_iter = 10000,
  rho_scale = 1,
  tolerance = 1e-04,
  linear_solver = c("kalman_filter", "sparse_qr"),
  space_tolerance_ratio = sqrt(.Machine$double.eps),
  ...
)
}
\arguments{
\item{max_iter}{Integer. Maximum number of iterations per \code{lambda} value.}

\item{rho_scale}{Double. The ratio of \code{lambda} divided by the augmented
Lagrangian penalty parameter \eqn{\rho}.}

\item{tolerance}{Double. The convergence tolerance for the ADMM algorithm.}

\item{linear_solver}{Integer. Solver for the linear system in ADMM when
k > 1: \code{kalman_filter} for Kalman filter or \code{sparse_qr} for sparse QR decomposition.}

\item{space_tolerance_ratio}{Double. The tolerance ratio of to detect that the signal is equally spaced.}

\item{...}{not used}
}
\value{
an object of class \code{admm_control}
}
\description{
These arguments are used to handle the internals of the ADMM algorithm
used to solve the trendfilter optimization problem.
}
\examples{
admm_control_list()
admm_control_list(max_iter = 10L)
admm_control_list(tolerance = 1e-8)
}
