% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-trendfilter.R
\name{plot.trendfilter}
\alias{plot.trendfilter}
\title{Plot estimated values from a \code{trendfilter} object}
\usage{
\method{plot}{trendfilter}(x, lambda = NULL, ...)
}
\arguments{
\item{x}{output of the function \code{\link[=trendfilter]{trendfilter()}} of class \code{trendfilter}}

\item{lambda}{select which estimates to plot. If not provided,
all estimates are plotted.}

\item{...}{Not used.}
}
\description{
Produces a figure showing some or all estimated values for different
values of the penalty. The result is a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}. Additional user
modifications can be added as desired.
}
\examples{
x <- 1:100 / 101 * 2 * pi
y <- sin(x) + .2 * rnorm(100)
out <- trendfilter(y, x, lambda = c(12, 1.2, .12))
plot(out)
}
